fix: viewport rendering — tonemap fullscreen + texture pointer truncation
Critical fixes for Metal framegraph viewport display:

1. GetViewportColorTexID: static_cast<uint32_t> → removed cast
   uint32_t truncated 64-bit Metal texture pointers (bit 32+ lost),
   causing ImGui::Image to crash with invalid texture address.

2. tonemap.slang: vertex shader now uses SV_VertexID to generate
   fullscreen triangle without a vertex buffer. Previous version expected
   Float2 vertex input but Execute called drawPrimitives without binding
   any VBO → GPU read garbage positions → black output.

3. RenderSystem: removed tonemapDesc.VertexLayout (empty = no VBO).

4. TextureMTLFull: storageMode Private→Shared for color render targets
   (ImGui fragment shader needs to sample them in the screen pass).

5. GetTextureNativeID: added validation logging.

+++ CHEngine/src/CHEngine/Render/RenderFacade.cpp
@@ -318,7 +318,7 @@ namespace CHEngine
         if (!s_ViewportOutputTex.IsValid()) return 0;
         IRenderFactory* f = GetFactory();
         if (!f) return 0;
-        return static_cast<uint32_t>(f->GetTextureNativeID(s_ViewportOutputTex));
+        return f->GetTextureNativeID(s_ViewportOutputTex);
     }
 
     // ── Default mesh shader ───────────────────────────────────────────────────
